+++ RoughSets/ApproxReductBoostingCV/Program.cs
@@ -59,7 +59,7 @@ namespace ApproxReductBoostingCV
                     ParameterValueCollection<WeightingSchema>.CreateFromElements("WeightingSchama", WeightingSchema.Majority),
                     ParameterValueCollection<bool>.CreateFromElements("CheckEnsembleErrorDuringTraining", false),
                     ParameterValueCollection<UpdateWeightsDelegate>.CreateFromElements("SetWeights", ReductEnsembleBoostingGenerator.UpdateWeightsAdaBoost_All),
-                    //ParameterValueCollection<int>.CreateFromElements("MinLenght", (int) System.Math.Floor(System.Math.Log((decimal)numOfAttr + 1.0M, 2.0)))
+                    //ParameterValueCollection<int>.CreateFromElements("MinLenght", (int) System.Math.Floor(System.Math.Log((double)numOfAttr + 1.0M, 2.0)))
                     //ParameterValueCollection<int>.CreateFromElements("MinLenght", 1)
                     new ParameterNumericRange<int>(ReductGeneratorParamHelper.Epsilon, 0, 50, 5)
                 }
@@ -152,12 +152,12 @@ namespace ApproxReductBoostingCV
                     }
 
                     parms.SetParameter(ReductGeneratorParamHelper.FactoryKey, factoryKey);
-                    parms.SetParameter(ReductGeneratorParamHelper.IdentificationType, (Func<long, IReduct, EquivalenceClass, decimal>)RuleQuality.ConfidenceW);
-                    parms.SetParameter(ReductGeneratorParamHelper.VoteType, (Func<long, IReduct, EquivalenceClass, decimal>)RuleQuality.ConfidenceW);
+                    parms.SetParameter(ReductGeneratorParamHelper.IdentificationType, (Func<long, IReduct, EquivalenceClass, double>)RuleQuality.ConfidenceW);
+                    parms.SetParameter(ReductGeneratorParamHelper.VoteType, (Func<long, IReduct, EquivalenceClass, double>)RuleQuality.ConfidenceW);
                     parms.SetParameter(ReductGeneratorParamHelper.NumberOfReductsInWeakClassifier, 1);
                     parms.SetParameter(ReductGeneratorParamHelper.MaxIterations, iter);
                     parms.SetParameter(ReductGeneratorParamHelper.UpdateWeights, updateWeights);
-                    parms.SetParameter(ReductGeneratorParamHelper.Epsilon, (decimal)epsilon / 100.0M);
+                    parms.SetParameter(ReductGeneratorParamHelper.Epsilon, epsilon / 100.0);
                     parms.SetParameter(ReductGeneratorParamHelper.WeightGenerator, weightGenerator);
                     parms.SetParameter(ReductGeneratorParamHelper.CheckEnsembleErrorDuringTraining, checkEnsembleErrorDuringTraining);
