+++ RoughSets/Infovision.Datamining.RoughSet/IReductGenerator.cs
@@ -5,7 +5,7 @@ namespace Infovision.Datamining.Roughset
 {
     public interface IReductGenerator
     {
-        decimal Epsilon { get; set; }
+        double Epsilon { get; set; }
         long ReductGenerationTime { get; }
 
         void InitFromArgs(Args args);
@@ -16,6 +16,6 @@ namespace Infovision.Datamining.Roughset
 
         IReductStoreCollection GetReductStoreCollection(int numberOfEnsembles = Int32.MaxValue);
 
-        IReduct CreateReduct(int[] permutation, decimal epsilon, decimal[] weights, IReductStore reductStore = null, IReductStoreCollection reductStoreCollection = null);
+        IReduct CreateReduct(int[] permutation, double epsilon, double[] weights, IReductStore reductStore = null, IReductStoreCollection reductStoreCollection = null);
     }
 }
