+++ RoughSets/Infovision.Datamining.RoughSet/ReductEnsembleBoostingVarEpsWithAttributeDiversity.cs
@@ -16,7 +16,7 @@ namespace Infovision.Datamining.Roughset
         {
         }
 
-        public override IReduct GetNextReduct(decimal[] weights)
+        public override IReduct GetNextReduct(double[] weights)
         {
             Permutation permutation = new PermutationGeneratorEnsemble(this.DataStore, this.GetReductGroups()).Generate(1)[0];
             return this.CreateReduct(permutation.ToArray(), this.Epsilon, weights);
