+++ RoughSets/Infovision.Datamining.Roughset.UnitTests/ReductDiscernibilityMatrixTest.cs
@@ -52,7 +52,7 @@ namespace Infovision.Datamining.Roughset.UnitTests
                     {
                         foreach (IReduct red in rs)
                         {
-                            decimal quality = InformationMeasureBase.Construct(InformationMeasureType.ObjectWeights).Calc(red);
+                            double quality = InformationMeasureBase.Construct(InformationMeasureType.ObjectWeights).Calc(red);
                             resultFile.WriteLine("{0} : {1}", red.ToString(), quality);
                         }
                     }
